+++ src/bindings/lua/Makefile.am
@@ -56,7 +56,6 @@ luamod_ldflags = \
 	$(san_ld_zdef_flag)
 
 luamod_libadd = \
-	$(top_builddir)/src/modules/kvs/libflux-kvs.la \
 	$(top_builddir)/src/modules/libkz/libkz.la \
 	$(top_builddir)/src/common/libflux-internal.la \
 	$(top_builddir)/src/common/libflux-core.la \
