+++ src/bindings/lua/kvs-lua.c
@@ -33,9 +33,9 @@
 #include <lua.h>
 #include <lauxlib.h>
 
-#include "flux/core.h"
+#include <flux/core.h>
 #include "src/common/libcompat/compat.h"
-#include "src/modules/kvs/kvs_deprecated.h"
+#include "src/common/libkvs/kvs_deprecated.h"
 
 #include "json-lua.h"
 #include "lutil.h"
