+++ src/common/libkvs/Makefile.am
@@ -1,0 +1,53 @@
+AM_CFLAGS = \
+	@GCCWARN@ \
+	$(CODE_COVERAGE_CFLAGS)
+
+AM_LDFLAGS = \
+	$(CODE_COVERAGE_LDFLAGS)
+
+AM_CPPFLAGS = \
+	-I$(top_srcdir) -I$(top_srcdir)/src/include \
+	$(ZMQ_CFLAGS)
+
+noinst_LTLIBRARIES = libkvs.la
+
+libkvs_la_SOURCES = \
+	kvs.c \
+	proto.c \
+	proto.h \
+	json_dirent.c \
+	json_dirent.h \
+	kvs_deprecated.h
+
+fluxinclude_HEADERS = \
+	kvs.h
+
+TESTS = \
+	test_proto.t \
+	test_dirent.t
+
+check_PROGRAMS = \
+	$(TESTS)
+
+TEST_EXTENSIONS = .t
+T_LOG_DRIVER = env AM_TAP_AWK='$(AWK)' $(SHELL) \
+	$(top_srcdir)/config/tap-driver.sh
+
+test_ldadd = \
+	$(top_builddir)/src/common/libflux-internal.la \
+	$(top_builddir)/src/common/libkvs/proto.o \
+	$(top_builddir)/src/common/libkvs/json_dirent.o \
+	$(top_builddir)/src/common/libflux-core.la \
+	$(top_builddir)/src/common/libtap/libtap.la
+
+test_cppflags = \
+	$(AM_CPPFLAGS) \
+	-I$(top_srcdir)/src/common/libtap
+
+test_proto_t_SOURCES = test/proto.c
+test_proto_t_CPPFLAGS = $(test_cppflags)
+test_proto_t_LDADD = $(test_ldadd)
+
+test_dirent_t_SOURCES = test/dirent.c
+test_dirent_t_CPPFLAGS = $(test_cppflags)
+test_dirent_t_LDADD = $(test_ldadd)
