+++ src/common/libkvs/json_dirent.c
