+++ src/include/flux/core.h
@@ -6,7 +6,7 @@
 
 #include "src/common/libflux/flux.h"
 
-#include "src/modules/kvs/kvs.h"
+#include "src/common/libkvs/kvs.h"
 #include "src/modules/libjsc/jstatctl.h"
 
 #endif /* FLUX_CORE_H */
