+++ src/modules/aggregator/Makefile.am
@@ -18,5 +18,4 @@ aggregator_la_SOURCES = aggregator.c
 aggregator_la_LDFLAGS = $(fluxmod_ldflags) -module
 aggregator_la_LIBADD = $(top_builddir)/src/common/libflux-internal.la \
 		 $(top_builddir)/src/common/libflux-core.la \
-		 $(top_builddir)/src/modules/kvs/libflux-kvs.la \
 		 $(ZMQ_LIBS)
