+++ src/modules/kvs/test/lookup.c
@@ -5,11 +5,11 @@
 
 #include "src/common/libutil/shortjson.h"
 #include "src/common/libtap/tap.h"
+#include "src/common/libkvs/proto.h"
+#include "src/common/libkvs/json_dirent.h"
 #include "src/modules/kvs/waitqueue.h"
 #include "src/modules/kvs/cache.h"
-#include "src/modules/kvs/proto.h"
 #include "src/modules/kvs/lookup.h"
-#include "src/modules/kvs/json_dirent.h"
 #include "src/modules/kvs/json_util.h"
 
 void basic_api (void)
