+++ src/modules/resource-hwloc/Makefile.am
@@ -19,6 +19,5 @@ resource_hwloc_la_CFLAGS = $(AM_CFLAGS) $(HWLOC_CFLAGS)
 resource_hwloc_la_LDFLAGS = $(fluxmod_ldflags) -module
 resource_hwloc_la_LIBADD = $(top_builddir)/src/common/libflux-core.la \
 			   $(top_builddir)/src/common/libflux-internal.la \
-			   $(top_builddir)/src/modules/kvs/libflux-kvs.la \
 			   $(LIBMUNGE) $(ZMQ_LIBS) $(LIBPTHREAD) \
 			   $(LIBUTIL) $(HWLOC_LIBS)
