+++ src/modules/wreck/Makefile.am
@@ -28,8 +28,7 @@ fluxlibexec_PROGRAMS = \
 	wrexecd
 
 fluxmod_libadd = $(top_builddir)/src/common/libflux-core.la \
-		 $(top_builddir)/src/common/libflux-internal.la \
-		 $(top_builddir)/src/modules/kvs/libflux-kvs.la
+		 $(top_builddir)/src/common/libflux-internal.la
 
 job_la_SOURCES = job.c
 job_la_LDFLAGS = $(AM_LDFLAGS) $(fluxmod_ldflags) -module
@@ -45,8 +44,7 @@ wrexecd_libs = \
 	$(top_builddir)/src/common/libflux-internal.la \
 	$(top_builddir)/src/common/libflux-core.la \
 	$(top_builddir)/src/common/libflux-optparse.la \
-	$(top_builddir)/src/modules/libkz/libkz.la \
-	$(top_builddir)/src/modules/kvs/libflux-kvs.la
+	$(top_builddir)/src/modules/libkz/libkz.la
 
 wrexecd_LDADD = \
 	$(wrexecd_libs) \
