+++ src/modules/wreck/wrexecd.c
@@ -51,8 +51,6 @@
 #include "src/common/libutil/shortjson.h"
 #include "src/common/libsubprocess/zio.h"
 #include "src/common/libpmi/simple_server.h"
-
-#include "src/modules/kvs/kvs.h"
 #include "src/modules/libkz/kz.h"
 
 #include "luastack.h"
